requirements.txt
audioread==2.1.4
cycler==0.10.0
Cython==0.25.1
decorator==4.0.10
funcsigs==1.0.2
h5py==2.6.0
joblib==0.10.3
Keras==1.1.2
librosa==0.4.3
matplotlib==1.5.3
mock==2.0.0
numpy==1.11.2
pbr==1.10.0
protobuf==3.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
resampy==0.1.4
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==0.12.0rc0
Theano==0.8.2